Allow deferring execution in runOnEventLoopAndWait
Default behavior for the runOn* methods is to execute the operation
immediately if the calling thread is driving the event loop (or,
equivalently, if there is no event loop in progress). This patch adds
an optional `defer` parameter that prevents this behavior. This
allows a caller to synchronize with execution of the event loop, e.g.
to block until a launched loop thread has begun executing.

Invoking runOnEventLoopAndWait(..., /*defer=*/ true) from the loop
thread will lead to deadlock.