Make test building optional
Adds an option `BUILD_TESTS` to the cmake configuration, allowing the
test targets to be bypassed. This also makes the gtest submodule
optional, which improves life when releasing.

Alternative: don't use a gtest submodule, but instead download like a
normal external project. Probably that would be a better strategy
going forward.